Organized project
Changed project folder organization and moved kosmickong.java to game folder from kosmickong folder, which i deleted. Updated main class in pom.xml. There is a game.java class in game folder which the kosmickong.java will act as instead, but I left the game.java there with the javadoc comments as a reference to things we might need to add going forward.