Working movement component
Movement component now works by taking the input from the inputhandler and updating the position of the entity accordingly. Still need to add logic for jumping in the movementcomponent. Currently the movementcomponent contains the player movement strategy in the update method which can later be changed to calling the entity's movement strategy's move method. The render component should be fully functional but I need to connect the homescreen and render component by using the homescreen to paint the graphics onto the UI