entity renders on screen and moves
created game panel to be a container for the entities which is held by the homescreen. Added list of entities with getter method and method in sprite class to load the image from a path. Added some basic functionality to the playermovementstrategy class to make sure it will properly connect dynamically with the movementcomponent, accept input, and render the changes on the screen, but will need to be continually updated as things like collision detection start.  Currently some test variables in the main class  which just spawns a player in the center of the screen which can only go left and right for now.